Refactor things from `test_type_promotion.py` to `dtype_helpers.py`
